Input validators and Network builders

CommonOPF.check_busses! — Function

check_busses!(busses::AbstractVector{<:AbstractBus}) = nothing

The default action after build_busses.

check_busses!(loads::AbstractVector{Load})

Remove (and warn about it) if any Load have no way to define the loads

CommonOPF.check_edges! — Function

check_edges!(edges::AbstractVector{<:AbstractEdge}) = nothing

The default action after build_edges.

check_edges!(conductors::AbstractVector{Conductor})

if all phases are missing then - warn_singlephase_conductors_and_copy_templates(conductors) else - validate_multiphase_edges!(conductors)

check_edges!(transformers::AbstractVector{Transformer})

fill in rmatrix and xmatrix if phases is not missing. For now assuming zero mutual impedances.

check_edges!(regulators::AbstractVector{VoltageRegulator})

Warn if not missing both vregpu and turnratio and call validatemultiphaseedges! if any phases are not missing.

CommonOPF.fill_node_attributes! — Function

fill_node_attributes!(g::MetaGraphsNext.AbstractGraph, vals::AbstractVector{<:AbstractBus})

For each concrete bus in vals store a dict of key,val pairs for the bus attributes in a symbol key like :Load for CommonOPF.Load TODO just store the type