Edge Impedances and Admittances

CommonOPF exports methods to get the real-valued resistance and reactance of Network edges, as well as the complex-valued impedance. If the Network is SinglePhase then scalar values are returned; if the Network is MultiPhase then matrix values are returned.

Impedance

CommonOPF.rij — Function

rij(i::AbstractString, j::AbstractString, net::Network)

resistance(net[(i,j)])

Resistance of edge i-j

rij(i::AbstractString, j::AbstractString, net:::Network{SinglePhase})

Scalar resistance of edge i-j

CommonOPF.rij_per_unit — Function

rij_per_unit(i::AbstractString, j::AbstractString, net::Network)

resistance(net[(i,j)]) / net.Zbase

Resistance of edge i-j normalized by net.Zbase

CommonOPF.xij — Function

xij(i::AbstractString, j::AbstractString, net::Network)

Reactance of edge i-j

xij(i::AbstractString, j::AbstractString, net:::Network{SinglePhase})

Scalar reactance of edge i-j

CommonOPF.xij_per_unit — Function

xij_per_unit(i::AbstractString, j::AbstractString, net::Network)

Reactance of edge i-j normalized by net.Zbase

CommonOPF.zij — Function

zij(i::AbstractString, j::AbstractString, net::Network)

Impedance matrix of edge (i,j)

CommonOPF.zij_per_unit — Function

zijperunit(i::AbstractString, j::AbstractString, net::Network)

Impedance matrix of edge (i,j) in per-unit (normalized with net.Zbase)

Admittance

CommonOPF.gij — Function

gij(i::AbstractString, j::AbstractString, net::Network)

conductance of edge i-j

gij(i::AbstractString, j::AbstractString, net:::Network{SinglePhase})

Scalar conductance of edge i-j

CommonOPF.gij_per_unit — Function

gij_per_unit(i::AbstractString, j::AbstractString, net::Network)

conductance of edge i-j normalized by net.Zbase

CommonOPF.bij — Function

bij(i::AbstractString, j::AbstractString, net::Network)

susceptance(net[(i,j)])

susceptance of edge i-j

bij(i::AbstractString, j::AbstractString, net:::Network{SinglePhase})

Scalar susceptance of edge i-j

CommonOPF.bij_per_unit — Function

bij_per_unit(i::AbstractString, j::AbstractString, net::Network)

susceptance(net[(i,j)]) * net.Zbase

susceptance of edge i-j normalized by net.Zbase

CommonOPF.yij — Function

yij(i::AbstractString, j::AbstractString, net::Network)

admittance matrix of edge (i,j)

yij(i::AbstractString, j::AbstractString, net::Network{SinglePhase})

admittance of edge (i,j)

CommonOPF.yij_per_unit — Function

yij_per_unit(i::AbstractString, j::AbstractString, net::Network{SinglePhase})

admittance of edge (i,j) in per-unit (multiplied with net.Zbase)

yij_per_unit(i::AbstractString, j::AbstractString, net::Network)

admittance matrix of edge (i,j) in per-unit (multiplied with net.Zbase)

Bus Admittance Matrix

The convention in CommonOPF is upper case Y for the bus admittance matrix methods:

CommonOPF.Yij — Function

Yij(
    i::AbstractString, 
    j::AbstractString, 
    net::CommonOPF.Network
)::Union{ComplexF64, Matrix{ComplexF64}}

Returns either:

entry of bus admittance matrix at i,j for single phase networks or

3x3 sub-matrix of bus admittance matrix for the phases connecting busses i and j

CommonOPF.Yij_per_unit — Function

Yij_per_unit(
    i::AbstractString, 
    j::AbstractString, 
    net::CommonOPF.Network
)::Union{ComplexF64, Matrix{ComplexF64}}

Returns either:

entry of bus admittance matrix at i,j for single phase networks or

3x3 sub-matrix of bus admittance matrix for the phases connecting busses i and j

(multiplied with net.Zbase)

CommonOPF.Ysparse — Function

function Ysparse(net::CommonOPF.Network)::SparseArrays.SparseMatrixCSC

Returns a Symmetric view of sparse upper triangular matrix

Shunt Admittance

Network specification yaml files can include bus admittance values like:

ShuntAdmittance:
  - bus: 1
    g: 0
    b: 0.0011

The admittances will be stored in the Network bus data and can be retrieved using the yj methods:

CommonOPF.yj — Function

yj(j::AbstractString, net::Network{SinglePhase})::ComplexF64

Shunt admittance of bus j

yj(j::AbstractString, net::Network{MultiPhase})::Matrix{ComplexF64}

Shunt admittance of bus j