Symmetrical Mutliphase Conductors

Often we only have the zero and positive sequence impedances of conductors. In these cases we construct the phase impedance matrix as:

\[z_{abc} = \begin{bmatrix} z_s & z_m & z_m \\ z_m & z_s & z_m \\ z_m & z_m & z_s \end{bmatrix}\]

where

\[\begin{aligned} z_s &= \frac{1}{3} z_0 + \frac{2}{3} z_1 \\ z_m &= \frac{1}{3} (z_0- z_1) \end{aligned} \]

Kron Reduction

CommonOPF.kron_reduce — Function

kron_reduce(M::AbstractMatrix)::Matrix

Given a 4x4 matrix remove the 4th row and column to create a 3x3 matrix.

The new values in the 3x3 are:

M_new[i,j] = M[i,j] - M[i,4] *  M[4,j] / M[4,4]  

where i and j are in the Set((1,2,3)).