Graphs

Methods for using/analyzing the network model as a graph

CommonOPF.make_graph — Function

make_graph(edges::AbstractVector{<:AbstractEdge};  directed::Union{Bool,Missing}=missing)

return MetaGraph made up of the edges

Also the graph[:intbusmap] is created with the dicts for bus => int and int => bus (because Graphs.jl only works with integer nodes)

julia> g["13", :bus]
10

julia> g[13, :bus]
"24"

julia> get_prop(g, :int_bus_map)[13]
"24"

CommonOPF.leaf_busses — Function

leaf_busses(net::Network)

returns Vector{String} containing all of the leaf busses in net.graph

IO methods

Network

CommonOPF.remove_bus! — Method

remove_bus!(j::String, net::Network{SinglePhase})

Remove bus j in the line i->j->k from the model by making an equivalent line from busses i->k

CommonOPF.reduce_tree! — Method

reduce_tree!(net::Network{SinglePhase})

combine any line sets with intermediate busses that have indegree == outdegree == 1 and is not a load bus into a single line

See remove_bus! for how the two lines are combined.

CommonOPF.trim_tree! — Function

trim_tree!(net::Network)

Trim any branches that have empty busses, i.e. remove the branches that have no loads or DER.

CommonOPF.trim_tree_once! — Function

trim_tree_once!(net::Network)

A support function for trim_tree!, trim_tree_once! removes all the empty leaf busses. When trimming the tree sometimes new leafs are created. So trim_tree! loops over trim_tree_once!.

Edge Impedances

CommonOPF.resistance — Function

resistance(c::Conductor)

resistance_per_length(c) * c.length

The absolute resistance of the conductor (in the units provided by the user)

CommonOPF.resistance_per_length — Function

resistance_per_length(c::Conductor)

if ismissing(c.phases)  # single phase
    return c.r1
end
return c.rmatrix

CommonOPF.rij — Function

rij(i::AbstractString, j::AbstractString, net::Network{SinglePhase})

resistance(net[(i,j)])

Resistance of edge i-j

rij(i::AbstractString, j::AbstractString, net::Network{MultiPhase})

Resistance of edge i-j

CommonOPF.rij_per_unit — Function

rij_per_unit(i::AbstractString, j::AbstractString, net::Network{SinglePhase})

resistance(net[(i,j)]) / net.Zbase

Resistance of edge i-j normalized by net.Zbase

CommonOPF.xij — Function

xij(i::AbstractString, j::AbstractString, net::Network{MultiPhase})

Reactance of edge i-j

CommonOPF.zij — Function

function zij(i::AbstractString, j::AbstractString, net::Network{SinglePhase})::Tuple{Real, Real}

Impedance for single phase models.

Returns (r1, x1) * length / net.Zbase for the Conductor at net[(i, j)].

TODO convert impedance methods to dispatch on edge type TODO MultiPhase