Variables for Mathematical Programs

CommonOPF provides standard variables names so that the packages that build upon CommonOPF can leverage common results getters and VariableBounds.

Variable Containers

CommonOPF provides variable containers to standardize indexing across OPF models. The order of indexing is:

var_symbol like :v

bus_or_edge

time_step

phase in [1, 2, 3] (if MultiPhase model)

Single phase models

CommonOPF.add_time_vector_variables! — Function

add_time_vector_variables!(
    m::JuMP.AbstractModel, 
    net::Network{SinglePhase}, 
    var_symbol::Symbol, 
    indices::AbstractVector{T} 
) where {T}

Add variables to the m.obj_dict indexed on:

var_symbol like :v

indices, typically a Vector{String} for bus indices or a Vector{Tuple{String, String}} for edge indices (like ("bus1", "bus2"))

timesteps in 1:net.Ntimesteps

For example, accessing the edge variable :sij one edge ("bus1", "bus2") voltage variable at time step 5 looks like:

m[:sij][("bus1", "bus2")][5]

Multiphase models

CommonOPF.multiphase_bus_variable_container — Function

multiphase_bus_variable_container(; default::DefaultType=missing)

Return a DefaultDict of DefaultDict with three key types for indexing variables on:

AbstractString for bus names

Int for time step

AbstractVecOrMat for vectors or matrices of phase variables

CommonOPF.multiphase_edge_variable_container — Function

multiphase_edge_variable_container(; default::DefaultType=missing)

Return a DefaultDict of DefaultDict with three key types for indexing variables on:

Tuple{String, String} for edge names

Int for time step

AbstractVecOrMat for vectors or matrices of phase variables

Variable Bounds

CommonOPF.VariableBounds — Type

struct VariableBounds

Limits for decision variables in mathematical programs. Upper and lower values can be specified for power, current, and voltage variables. The VariableBounds struct is attached to Network.bounds upon creation of the Network.

mutable struct VariableBounds
    s_upper_real::Union{Real, Missing}
    s_lower_real::Union{Real, Missing}
    s_upper_imag::Union{Real, Missing}
    s_lower_imag::Union{Real, Missing}
    
    v_upper_mag::Union{Real, Missing}
    v_lower_mag::Union{Real, Missing}

    i_upper_mag::Union{Real, Missing}
    i_lower_mag::Union{Real, Missing}
end

CommonOPF.VariableBounds — Method

function VariableBounds(ntwk::Dict)

Check for the keys of the VariableBounds struct in the ntwk dictionary. Any missing bounds are set to missing.

Documenting Variables

CommonOPF.VarInfo — Type

struct VarInfo
    symbol::Symbol
    description::String
    units::VarUnits
    dimensions::Tuple{Vararg{VarDimensions}}
end

Variable information for describing variables in sub-modules of CommonOPF. See also VarUnits, VarDimensions, and print_var_info.

CommonOPF.VarUnits — Type

@enum VarUnits begin
    AmpUnit
    VoltUnit
    TimeUnit
    RealPowerUnit
    ReactivePowerUnit
    ApparentPowerUnit
end

Units acceptable for CommonOPF variables.

CommonOPF.VarDimensions — Type

@enum VarDimensions begin
    BusDimension
    EdgeDimension
    TimeDimension
    PhaseDimension
end

Dimensions for specifying variable indexes in CommonOPF variables.

CommonOPF.print_var_info — Function

print_var_info(net::Network)

Print the variable info in net.var_info in a table like:

julia> print_var_info(net)
┌────────┬───────────────────────────┬──────────┬────────────────────┐
│ symbol │               description │    units │         dimensions │
├────────┼───────────────────────────┼──────────┼────────────────────┤
│  vsqrd │ voltage magnitude squared │ VoltUnit │ (Bus, Time, Phase) │
└────────┴───────────────────────────┴──────────┴────────────────────┘