Graphs

Methods for using/analyzing the network model as a graph

CommonOPF.make_graph — Function

make_graph(edges::AbstractVector)

return SimpleDiGraph by inferring busses from edges with the dicts for bus => int and int => bus (because Graphs.jl only works with integer nodes)

julia> g["13", :bus]
10

julia> get_prop(g, :bus_int_map)["13"]
10

julia> g[13, :bus]
"24"

julia> get_prop(g, :int_bus_map)[13]
"24"

make_graph(busses::AbstractVector{String}, edges::AbstractVector)

return SimpleDiGraph with the dicts for bus => int and int => bus (because Graphs.jl only works with integer nodes)

julia> g["13", :bus]
10

julia> get_prop(g, :bus_int_map)["13"]
10

julia> g[13, :bus]
"24"

julia> get_prop(g, :int_bus_map)[13]
"24"

CommonOPF.leaf_busses — Function

leaf_busses(p::Inputs)

returns Vector{String} containing all of the leaf busses in p.busses

IO methods

Inputs

Types

Utils

CommonOPF.trim_tree! — Function

trim_tree!(p::Inputs)

Trim any branches that do not contain load busses.

CommonOPF.trim_tree_once! — Function

trim_tree_once!(p::Inputs)

A support function for trim_tree!. When trimming the tree sometimes new leafs are created. So trim_tree! loops over trim_tree_once!.